don't decode on Python 2
In Python 3, strings are no longer just generic bundles of bytes.  A decoding
step was introduced in (36:8e47cd330f57) to convert bytes to unicode on Python
3 (as this was the least invasive way to make the rest of the "string-based"
codebase work).

This appeared to work for Python 2.  However, in the presence of high-bit
characters in the input stream, this causes errors due to decoding.  This has
been reported as Issues 23 and 26.

Specifically, the default string encoding is 'ascii' and the default error
handling mode of 'strict'.   While this doesn't handle all encoding completely
correctly, it passes through data unmodified and isn't really less correct than
the current behavior.

In lieu of fully compliant encoding handling, this should smooth over some
spurious errors for messages with binary content.