Added Python 3.4 minimum requirement to setup
stomp.py 5.0.0 removed support for Python 2.7, but that is not properly
indicated in the package metadata. Therefore, pip installs stomp.py 5.0.0
and higher happily on Python 2.7.

To constrain the Python version, the 'python_requires' argument for
'setuptools.setup()' needs to be specified with the desired version
constraints. The 'classifiers' argument of 'setuptools.setup()' is
not used for that.

This change adds 'python_requires=">=3.4"' in order to constrain the
Python version accordingly.

Solves issue #279.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>