fix: Reconnect failure
The "notified on disconnect" flag prevents double disconnects but once
it is set it is not being reset. So it does prevent double disconnect
event on first disconnect but on second disconnect it doesn't fire
"disconnected" event at all. This fix resets the "notified on
disconnect" after skipping the "disconnected" event.

closes issue #294