remove unnecessary certificate expiration check and re-enable test
When connecting to a server with an expired certificate, the ssl library will raise:

ssl.SSLCertVerificationError: [SSL: CERTIFICATE_VERIFY_FAILED] certificate verify failed: certificate has expired

This makes check_ssl_certificate() unnecessary.

Note that using ssl.get_server_certificate() in that function creates a second
connection to the server, independent of the SSL parameters used for the first
connection like cacerts and client certs. This can result in errors, particularly in
cases where the server is configured to require client certificate authentication.
Removing the unnecessary additional check avoids the risk of errors from this second
connection.

This also removes the dependency on cryptography, which was only used in
check_ssl_certificate().