mock dns responses to make running tests easier and more efficient
By monkeypatching socket.getaddrinfo() in a couple of places, the requirement to add
entries to the hosts file (which usually requires root access) is removed. It also
avoid long timeouts in dual-stack IPv4/IPv6 environments where the tests try to connect
to the IPv6 address for localhost first, but the test service is only listening on the
IPv4 address.

Also update test_alternate_hosts() to wait for a successful connection, and fix a
logic error in StubStompServer where get_next_frame() was returning an empty string but the
caller was testing for None.