Use single thread for IO when using TLS
Previously reads always happened in a separate thread (the "receiver
thread") while writes happened in the main thread. This is fine for
plain TCP sockets for the TLS the situation is less clear. Underlying
SSL_read/SSL_write functions are not thread-safe however Python's ssl
module might add some or full thread-safety. Either way, in
https://github.com/jasonrbriggs/stomp.py/issues/450 two different errors
where observed when connecting to a TLS server:

1. Random ConnectFailedException. According to debug logs, the "STOMP"
   message was sent but it was seemingly never received by the server
   and the server eventually closed the connection.

2. Sending the "STOMP" message sometimes resulted in
   ssl.SSLEOFError: EOF occurred in violation of protocol (_ssl.c:2427)"

To ensure we don't run into thread-safety issues we now do both reads
and writes in the same thread for TLS. Since we need to be able to
receive and send messages concurrently we also switch to a non-blocking
socket.

Non-TLS connections are not affected by this change, they read and write
in different threads like before.